# core-lending-asset-finance-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/usage-records/{recordId}:
    get:
      tags:
      - UsageRecord
      summary: Get a usage record by ID
      description: Retrieve a specific usage record by its unique identifier
      operationId: getById_3
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the usage record
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Usage record found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageRecordDTO"
        "404":
          description: "Usage record, asset, or agreement not found"
        "500":
          description: Internal server error
    put:
      tags:
      - UsageRecord
      summary: Update an existing usage record
      description: Update all fields of an existing usage record
      operationId: update_3
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the usage record to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageRecordDTO"
        required: true
      responses:
        "200":
          description: Usage record updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageRecordDTO"
        "400":
          description: Invalid input data
        "404":
          description: "Usage record, asset, or agreement not found"
        "500":
          description: Internal server error
    delete:
      tags:
      - UsageRecord
      summary: Delete a usage record
      description: Permanently delete a usage record
      operationId: delete_3
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the usage record to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Usage record deleted successfully
        "404":
          description: "Usage record, asset, or agreement not found"
        "500":
          description: Internal server error
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/usage-records:
    get:
      tags:
      - UsageRecord
      summary: List/Search usage records for an asset
      description: "Retrieve a paginated list of usage records (mileage, hours, etc.)\
        \ for a specific asset"
      operationId: findAll_3
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: filterRequest
        in: query
        description: Filter criteria for searching usage records
        required: true
        schema:
          $ref: "#/components/schemas/FilterRequestUsageRecordDTO"
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved list of usage records
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginationResponseUsageRecordDTO"
        "400":
          description: Invalid filter request
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
    post:
      tags:
      - UsageRecord
      summary: Create a new usage record for an asset
      description: "Create a new usage record to track asset usage metrics (mileage,\
        \ operating hours, etc.)"
      operationId: create_3
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageRecordDTO"
        required: true
      responses:
        "200":
          description: Usage record created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageRecordDTO"
        "400":
          description: Invalid input data
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    UsageRecordDTO:
      required:
      - assetFinanceAssetId
      - usageDate
      type: object
      properties:
        usageRecordId:
          type: string
          format: uuid
          readOnly: true
        assetFinanceAssetId:
          type: string
          format: uuid
        usageDate:
          type: string
          format: date
        mileage:
          minimum: 0
          type: integer
          format: int32
        usageDetail:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: "The direction of sorting, either ASC or DESC."
          example: DESC
          default: DESC
      description: "Represents a pagination request for retrieving paginated results,\
        \ including page number, size, sort field, and direction."
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RangeObject"
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestUsageRecordDTO:
      required:
      - pagination
      type: object
      properties:
        filters:
          $ref: "#/components/schemas/UsageRecordDTO"
        rangeFilters:
          $ref: "#/components/schemas/RangeFilter"
        pagination:
          $ref: "#/components/schemas/PaginationRequest"
        options:
          $ref: "#/components/schemas/FilterOptions"
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseUsageRecordDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: "#/components/schemas/UsageRecordDTO"
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: "The current page number, typically zero-based."
          format: int32
      description: Represents a paginated response containing a list of items and
        pagination metadata.
...